+++ apps/web/src/app/organizations/vault/organization-vault.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from "@angular/core";
 
-import { VaultModule } from "../../vault.module";
+import { VaultModule } from "../../vault/shared/vault.module";
 
 import { OrganizationVaultRoutingModule } from "./organization-vault-routing.module";
 import { OrganizationVaultComponent } from "./organization-vault.component";
