+++ apps/web/src/app/shared/loose-components.module.ts
@@ -29,7 +29,6 @@ import { FrontendLayoutComponent } from "../layouts/frontend-layout.component";
 import { NavbarComponent } from "../layouts/navbar.component";
 import { UserLayoutComponent } from "../layouts/user-layout.component";
 import { VaultFilterModule } from "../modules/vault-filter/vault-filter.module";
-import { OrganizationBadgeModule } from "../modules/vault/modules/organization-badge/organization-badge.module";
 import { OrganizationCreateModule } from "../organizations/create/organization-create.module";
 import { OrganizationLayoutComponent } from "../organizations/layouts/organization-layout.component";
 import { BulkConfirmComponent as OrgBulkConfirmComponent } from "../organizations/manage/bulk/bulk-confirm.component";
@@ -158,6 +157,7 @@ import { BulkShareComponent } from "../vault/bulk-share.component";
 import { CiphersComponent } from "../vault/ciphers.component";
 import { CollectionsComponent } from "../vault/collections.component";
 import { FolderAddEditComponent } from "../vault/folder-add-edit.component";
+import { OrganizationBadgeModule } from "../vault/organization-badge/organization-badge.module";
 import { ShareComponent } from "../vault/share.component";
 
 // Please do not add to this list of declarations - we should refactor these into modules when doing so makes sense until there are none left.
