+++ apps/web/src/app/vault/individual-vault.component.html
