+++ apps/web/src/app/vault/individual-vault.module.ts
@@ -1,9 +1,8 @@
 import { NgModule } from "@angular/core";
 
-import { VaultModule } from "../../vault.module";
-
 import { IndividualVaultRoutingModule } from "./individual-vault-routing.module";
 import { IndividualVaultComponent } from "./individual-vault.component";
+import { VaultModule } from "./shared/vault.module";
 
 @NgModule({
   imports: [VaultModule, IndividualVaultRoutingModule],
