+++ apps/web/src/app/vault/organization-badge/organization-name-badge.component.ts
