+++ apps/web/src/app/vault/shared/vault.service.ts
